microservices/GCCONTENT/requirements.txt
confluent-kafka
kafka-python
biopython